chore: Fixup of #12 (#13)
* fix: change profile image handling

Change profile image handling process.
Do not limit file size but will resize to 250x250 pixels or less.
Special profile images (admin, no image) will also be saved.
Overwritten images will be deleted.
Let users set profile images during signup.

* fix: format and prune

Remove unnecessary comments and uncomment necessary code.
Refactor settings and code.

+++ backend/users/signals.py
@@ -50,5 +50,6 @@ def delete_profile(
     instance: AbstractUser,
     **kwargs,
 ):  # pylint: disable=W0613
-    profile: Profile = instance.profile  # type: ignore
-    profile.delete()
+    if hasattr(instance, "profile"):
+        profile: Profile = instance.profile
+        profile.delete()
